% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entrez.R
\name{efetch}
\alias{efetch}
\title{Function for retrieving PubMed search results
This function is for retrieving structured text from NCBI databases}
\usage{
efetch(query_string, db = "pubmed", type = "medline")
}
\arguments{
\item{query_string}{The query to used to retrieve results.}

\item{db}{Specifies what NCBI database to run th query against.}

\item{type}{Specifies the format results should be returned in i.e. medline}
}
\description{
Function for retrieving PubMed search results
This function is for retrieving structured text from NCBI databases
}
\examples{
efetch("pharmacy[AD]", "pubmed", "medline")
}

