% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entrez.R
\name{extractFields}
\alias{extractFields}
\title{Function for extracting certain fields from medline records.
This function take medline records and a field argument and returns a list of fields.}
\usage{
extractFields(recs, key, sep = FALSE)
}
\arguments{
\item{recs}{The list of medline records returned from medlineParser.}

\item{key}{The field to be extracted from medline records.}
}
\description{
Function for extracting certain fields from medline records.
This function take medline records and a field argument and returns a list of fields.
}

