% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/entrez.R
\name{medline_parser}
\alias{medline_parser}
\title{Function for parsing Medline files
This function is for parsing MEDLINE structured text i.e files or results from entrez_fetcher()}
\usage{
medline_parser(file_name)
}
\arguments{
\item{file_name}{Path to MEDLINE file or output from entrez_fetcher()}
}
\description{
Function for parsing Medline files
This function is for parsing MEDLINE structured text i.e files or results from entrez_fetcher()
}
\examples{
medline_parser(<path_to_medline_file>)
}
\keyword{NCBI}
\keyword{medline,}
\keyword{pubed,}

