% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entrez.R
\name{medlineDocumentTermMatrix}
\alias{medlineDocumentTermMatrix}
\title{Returns a document-term matrix using field specified by user}
\usage{
medlineDocumentTermMatrix(recs, key, label = "AU")
}
\description{
Returns a document-term matrix using field specified by user
}

