% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entrez.R
\name{medlineParser}
\alias{medlineParser}
\title{MEDLINE Parser}
\usage{
medlineParser(file_name)
}
\arguments{
\item{file_name}{Path to MEDLINE file or output from entrez_fetcher()}
}
\description{
MEDLINE Parser
}
\examples{
# If using MEDLINE file on local machine
results <- medlineParser("/home/user/Downloads/pubmed_results.txt")

# If using output from efetch()
search_results <- efetch("\\"university of new mexico\\"[AD] AND \\"pharmacy\\"[AD]")
results <- medlineParser(search_results)
}
\keyword{NCBI}
\keyword{medline,}
\keyword{pubed,}

